requirements.txt
feedparser
sentence-transformers
torch
numpy
scipy
requests
tqdm
click
typer
huggingface-hub
python-dotenv
fastapi
uvicorn
openai
PyPDF2
python-multipart
